move result table to report
