+++ src/offat/report/table.py
