+++ src/pyproject.toml
@@ -19,6 +19,7 @@ uvicorn = {extras = ["standard"], version = "^0.23.2", optional = true}
 rq = {version = "^1.15.1", optional = true}
 redis = {version = "^5.0.0", optional = true}
 python-dotenv = {version = "^1.0.0", optional = true}
+rich = "^13.7.0"
 
 [tool.poetry.group.dev.dependencies]
 pytest = "^7.4.0"
